+++ src/framework/src/Exception/ContextException.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Swoft\Exception;
+
+/**
+ * Class ContextException
+ * @since 2.0
+ */
+class ContextException extends \RuntimeException
+{
+
+}