+++ src/websocket-server/src/Exception/WsContextException.php
@@ -1,0 +1,12 @@
+<?php declare(strict_types=1);
+
+namespace Swoft\WebSocket\Server\Exception;
+
+/**
+ * Class WsContextException
+ * @since 2.0
+ */
+class WsContextException extends WsServerException
+{
+
+}